algorithm update
algorithm is updated based on the discussion we conducted on Tuesday. Since the server is not able to provide enough info for us to debug and test, I will not go forward. I am thinking about the thing that we can remove the button on the second page which is  currently sending locations to TTS, I think we dont need that anymore.